Groupe les pages profil et settings user et utilise une présentation similaire aux settings globales
