+++ client/src/locale/en/pages.js
@@ -14,27 +14,6 @@ export default {
         },
     },
 
-    'page-profile': {
-        'title': "Your profile",
-        'help': "If you change your email, pseudo or password, do not forget them before logging-out!",
-        'you-are-group': "You are: {group}.",
-        'edit-password': "Change your password",
-        'password-confirmation': "Password confirmation",
-        'password-confirmation-must-match': "Password and its confirmation must be identical.",
-        'password-modified': "Your password was modified.",
-        'saved': "Your profile was saved.",
-    },
-
-    'page-user-settings': {
-        'title': "Your settings",
-        'help': "The \"duration of a session\" is the time allowed before you're automatically logged out.",
-        'auth-token-validity-duration': "Max. duration of a session",
-        'interface': "Interface",
-        'language': "Language",
-        'hours': "hours",
-        'saved': "Settings saved.",
-    },
-
     'page-calendar': {
         'title': "Calendar",
         'help': (
@@ -408,6 +387,29 @@ export default {
             ),
             'public-calendar-url-reset-error': "An error occurred while regenerating the public calendar link, please try again.",
             'public-calendar-url-reset-success': "The public calendar link has been successfully re-generated! The new one is available above.",
+        },
+    },
+
+    'page-user-settings': {
+        'title': "Your settings",
+        'profile': {
+            'title': "Your profile",
+            'help': "If you change your email, pseudo or password, do not forget them before logging-out!",
+            'new-password': "New password",
+            'new-password-help': "Only fill this section if you want to change your password.",
+            'password-confirmation': "Password confirmation",
+            'password-confirmation-must-match': "Password and its confirmation must be identical.",
+            'password-modified': "Your password was modified.",
+            'saved': "Your informations have been successfully saved.",
+            'saved-with-password': "Your informations and your new password have been successfully saved.",
+        },
+        'interface': {
+            'title': "Interface",
+            'help': "The \"duration of a session\" is the time allowed before you're automatically logged out.",
+            'auth-token-validity-duration': "Max. duration of a session",
+            'language': "Language",
+            'hours': "hours",
+            'saved': "Settings saved.",
         },
     },
