+++ client/src/pages/Settings/User/Interface/index.js
@@ -2,13 +2,14 @@ import './index.scss';
 import Vue from 'vue';
 import Help from '@/components/Help';
 import FormField from '@/components/FormField';
+import Button from '@/components/Button';
 
 // @vue/component
 export default {
-    name: 'UserSettings',
+    name: 'InterfaceUserSettings',
     data() {
         return {
-            help: 'page-user-settings.help',
+            help: 'page-user-settings.interface.help',
             error: null,
             isLoading: true,
             langsOptions: [
@@ -68,7 +69,7 @@ export default {
             try {
                 const { data } = await this.$http.put(`users/${id}/settings`, this.settings);
                 this.settings = data;
-                this.help = { type: 'success', text: 'page-user-settings.saved' };
+                this.help = { type: 'success', text: 'page-user-settings.interface.saved' };
 
                 const userLocale = data.language.toLowerCase();
                 localStorage.setItem('userLocale', userLocale);
@@ -82,7 +83,7 @@ export default {
         },
 
         displayError(error) {
-            this.help = 'page-user-settings.help';
+            this.help = 'page-user-settings.interface.help';
             this.error = error;
 
             const { code, details } = error.response?.data?.error || { code: 0, details: {} };
@@ -104,51 +105,34 @@ export default {
         } = this;
 
         return (
-            <div class="content">
-                <div class="content__main-view UserSettings">
-                    <h3 class="UserSettings__title">{__('page-user-settings.interface')}</h3>
-                    <div class="UserSettings__content">
-                        <form class="Form" method="POST" onSubmit={handleSave}>
-                            <section class="Form__fieldset">
-                                <FormField
-                                    type="select"
-                                    options={langsOptions}
-                                    vModel={settings.language}
-                                    errors={errors.language}
-                                    name="language"
-                                    label="page-user-settings.language"
-                                />
-                                <FormField
-                                    type="number"
-                                    vModel={settings.auth_token_validity_duration}
-                                    errors={errors.auth_token_validity_duration}
-                                    name="auth_token_validity_duration"
-                                    label="page-user-settings.auth-token-validity-duration"
-                                    class="UserSettings__hours"
-                                    addon={__('hours')}
-                                />
-                            </section>
-                            <section class="Form__actions">
-                                <button class="Form__actions__save success" type="submit">
-                                    {__('save')}
-                                </button>
-                            </section>
-                        </form>
-                        <div class="UserSettings__extras">
-                            <Help message={help} error={error} isLoading={isLoading} />
-                            <div class="UserSettings__extras__buttons">
-                                <router-link to="/profile" custom>
-                                    {({ navigate }) => (
-                                        <button type="button" onClick={navigate} class="info">
-                                            <i class="fas fa-user-alt" />
-                                            {__('your-profile')}
-                                        </button>
-                                    )}
-                                </router-link>
-                            </div>
-                        </div>
-                    </div>
-                </div>
+            <div class="InterfaceUserSettings">
+                <Help message={help} error={error} isLoading={isLoading} />
+                <form class="InterfaceUserSettings__form" method="POST" onSubmit={handleSave}>
+                    <section class="InterfaceUserSettings__inputs">
+                        <FormField
+                            type="select"
+                            options={langsOptions}
+                            vModel={settings.language}
+                            errors={errors.language}
+                            name="language"
+                            label="page-user-settings.interface.language"
+                        />
+                        <FormField
+                            type="number"
+                            vModel={settings.auth_token_validity_duration}
+                            errors={errors.auth_token_validity_duration}
+                            name="auth_token_validity_duration"
+                            label="page-user-settings.interface.auth-token-validity-duration"
+                            class="InterfaceUserSettings__hours"
+                            addon={__('hours')}
+                        />
+                    </section>
+                    <section class="InterfaceUserSettings__actions">
+                        <Button icon="save" htmlType="submit" class="success">
+                            {__('save')}
+                        </Button>
+                    </section>
+                </form>
             </div>
         );
     },
