+++ client/src/pages/Settings/User/Interface/index.scss
@@ -1,0 +1,33 @@
+@use '~@/style/globals';
+
+.InterfaceUserSettings {
+    &__form {
+        max-width: globals.$form-max-width;
+    }
+
+    &__inputs {
+        margin-top: globals.$content-padding-large-vertical;
+
+        .FormField {
+            &__label {
+                flex-basis: 200px;
+            }
+
+            &__error__text {
+                padding-left: 200px;
+            }
+
+            + .FormField {
+                margin-top: globals.$content-padding-small-vertical;
+            }
+        }
+    }
+
+    &__hours .FormField__input {
+        max-width: 70px;
+    }
+
+    &__actions {
+        margin-top: globals.$content-padding-large-vertical;
+    }
+}
